GlobalProfilesController provides only profile selection
