--- src/main/resources/egl/client/controller/profile/global/GlobalProfilesController.fxml
@@ -1,66 +1,0 @@
-<?import egl.client.view.pane.CustomBorderPane?>
-<?import egl.client.view.text.LabeledTextField?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ListView?>
-<?import javafx.scene.control.Tab?>
-<?import javafx.scene.control.TabPane?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.text.Text?>
-<CustomBorderPane fx:controller="egl.client.controller.profile.global.GlobalProfilesController"
-                  xmlns:fx="http://javafx.com/fxml" margin="5">
-
-    <top>
-        <CustomBorderPane>
-            <right>
-                <Button fx:id="exitProfileButton" text="Выйти из профиля"/>
-            </right>
-        </CustomBorderPane>
-    </top>
-    <center>
-        <TabPane fx:id="activitiesTabPane" tabClosingPolicy="UNAVAILABLE">
-            <Tab fx:id="globalProfileTab" text="Глобальный профиль">
-                <GridPane fx:id="loginGridPane">
-                    <columnConstraints>
-                        <ColumnConstraints percentWidth="50"/>
-                        <ColumnConstraints percentWidth="50"/>
-                    </columnConstraints>
-                    <Text fx:id="loginInfoText" GridPane.rowIndex="0" GridPane.columnSpan="2" textAlignment="CENTER" textOrigin="CENTER"/>
-                    <LabeledTextField fx:id="loginTextField" label="Логин" GridPane.rowIndex="1" GridPane.columnIndex="0"/>
-                    <LabeledTextField fx:id="passwordTextField" label="Пароль" GridPane.rowIndex="2" GridPane.columnIndex="0"/>
-                    <Button fx:id="editProfileButton" GridPane.rowIndex="1" GridPane.columnIndex="1" visible="false" text="Изменить данные профиля"/>
-                    <Button fx:id="loginButton" GridPane.rowIndex="2" GridPane.columnIndex="1" text="Войти в профиль"/>
-                    <Text fx:id="errorText" GridPane.rowIndex="3" GridPane.columnSpan="2" textAlignment="CENTER" textOrigin="CENTER"/>
-                </GridPane>
-            </Tab>
-            <Tab fx:id="globalRatingsTab" text="Глобальные рейтинги">
-                <CustomBorderPane>
-                    <left>
-                        <ListView fx:id="topicsListView"/>
-                    </left>
-                    <top>
-                        <CustomBorderPane>
-                            <right>
-                                <Text fx:id="selectedTopicInfoText"/>
-                            </right>
-                        </CustomBorderPane>
-                    </top>
-                    <center>
-                        <CustomBorderPane>
-                            <center>
-                                <TabPane fx:id="taskRatingsTabPane" tabClosingPolicy="UNAVAILABLE"/>
-                            </center>
-                        </CustomBorderPane>
-                    </center>
-                </CustomBorderPane>
-            </Tab>
-        </TabPane>
-    </center>
-    <bottom>
-        <CustomBorderPane>
-            <right>
-                <Button fx:id="createProfileButton" text="Зарегистрировать профиль"/>
-            </right>
-        </CustomBorderPane>
-    </bottom>
-</CustomBorderPane>
